Everything is working!
We should probably have a way to import sequents or something, but as it stands this takes sequents and outputs trees. My next step will be to try to have everything typeset itself into something that looks like a tree when you ask it to. There's some cleanup to do and the code is fragile, but as far as a proof of concept I think this does exactly what we want it to do.